fix a bunch of warnings
git-svn-id: https://svn.code.sf.net/p/ec2drv/code/ec2drv/trunk@49 07914c0c-2d3e-0410-adf3-90c1a7ff5250

+++ src/ec2drv/devices.h
@@ -15,7 +15,7 @@ typedef struct
 	
 	uint16_t	internal_xram_size;
 	uint16_t	flash_size;
-	uint8_t		flash_sector_size;
+	uint16_t	flash_sector_size;
 	BOOL		tested;		// TRUE if ec2dev developers are happy the device support has been tested and is complete
 } EC2_DEVICE;
